Apply network update to app
