+++ packages/web-app/src/context/apolloClient.tsx
@@ -127,6 +127,7 @@ const client: Record<
   mumbai: mumbaiClient,
   arbitrum: undefined,
   'arbitrum-test': arbitrumTestClient,
+  unsupported: undefined,
 };
 
 /*************************************************
